refactor(api-client): modularize HTTP client for RN-ready web bootstrap (#156)
Split @bizcode/api-client into factory, errors, domain modules, and portal; bind VITE_API_URL from apps/web via initApiClientFromEnv. No import.meta in the package. Update trilingual local-dev docs and changelogs.

Closes #156
