+++ apps/web/src/main.tsx
@@ -3,7 +3,10 @@ import ReactDOM from 'react-dom/client'
 import App from './App'
 import './index.css'
 import './i18n/config'
+import { initApiClientFromEnv } from '@/lib/api-config'
 import { AuthProvider } from '@/contexts/AuthContext'
+
+initApiClientFromEnv()
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
