+++ packages/api-client/tsconfig.json
@@ -6,8 +6,7 @@
     "outDir": "dist",
     "declaration": true,
     "declarationMap": true,
-    "noEmit": false,
-    "types": ["vite/client"]
+    "noEmit": false
   },
   "include": ["src"],
   "references": [{ "path": "../types" }]
