+++ vitest.config.ts
@@ -46,8 +46,8 @@ export default defineConfig({
         'apps/web/src/types.ts',
         'apps/web/src/lib/plan-sync/index.ts',
         'apps/web/src/lib/api.ts',
+        'apps/web/src/lib/api-config.ts',
         'apps/web/src/lib/rbac.ts',
-        'apps/web/src/lib/portalApi.ts',
       ],
     },
   },
